+++ missense_kinase_toolkit/databases/mkt/databases/app/properties.py
