+++ missense_kinase_toolkit/databases/pyproject.toml
@@ -44,6 +44,7 @@ dependencies = [
     "upsetplot>=0.9.0",
     "xlrd>=2.0.1",
     "rdkit>=2023.9.6",
+    "webcolors>=24.11.1",
 ]
 
 [project.optional-dependencies]
